Remove non-pinned packages more thoroughly
This is easier explained with an example. Let's say I have AWSS3.jl and AWS.jl in my Project.toml, and I was working on AWS.jl so I un-pinned it so that I don't build it into my sysimage, because I want to be able to change versions (e.g. switch to my branch) without rebuilding it. Currently, we would `rm` AWS.jl from the Project before building the sysimage. But we would *not* rm AWSS3, which means that AWSS3 and *all of its dependencies* including AWS.jl would get baked into the sysimage. That means I would not be able to change versions of AWS.jl despite unpinning it.

Here, we use a Pkg command to remove the packages everywhere in the Manifest. That means if any package depends on a removed package, it is removed too. So in the example, we would not build either AWSS3 or AWS into the sysimage.

That means if you do not pin something very low in the stack, you might end up removing most things from your sysimage. But that's basically the cost of editing something low in the stack I think; it seems unavoidable.